run all 'non-admin' tests in one sandbox
