+++ __tests__/admins.ava.ts
@@ -7,30 +7,6 @@ export function tests(workspace: Workspace) {
   /**
    * add admins
    */
-  workspace.test('add admins by non-admin', async (test, {alice, bob, multicall, root}) => {
-    let callError: any;
-    // bob isn't admin so he can't add admins
-    try {
-      // try catch bacause contract should panick
-      await bob.call(
-        multicall.accountId,
-        'admins_add',
-        {
-          account_ids: [bob.accountId]
-        },
-        {
-          gas: Gas.parse('20 Tgas'),
-          attachedDeposit: NEAR.from('1') // 1 yocto
-        }
-      );
-    } catch (error) { callError = getFunctionCallError(error) }
-    const admins: string[] = await multicall.view('get_admins', {})
-    test.true(
-      ( admins.includes(alice.accountId) && !admins.includes(bob.accountId) )
-      && ( callError.includes("must be admin to call this function") )
-    );
-    test.log(`admins: [${admins}]`);
-  });
   workspace.test('add admins by admin', async (test, {alice, bob, multicall, root}) => {
     // alice is admin so she can add bob
     await alice.call(
@@ -54,29 +30,7 @@ export function tests(workspace: Workspace) {
   /**
    * remove admins
    */
-  workspace.test('remove admins by non-admin', async (test, {alice, bob, multicall, root}) => {
-    let callError: any;
-    // bob isn't admin so he can't remove admins
-    try {
-      // try catch bacause contract should panick
-      await bob.call(
-        multicall.accountId,
-        'admins_remove',
-        { account_ids: [alice.accountId] },
-        {
-          gas: Gas.parse('20 Tgas'),
-          attachedDeposit: NEAR.from('1') // 1 yocto
-        }
-      );
-    } catch (error) { callError = getFunctionCallError(error) }
-    const admins: string[] = await multicall.view('get_admins', {})
-    test.true(
-      ( admins.includes(alice.accountId) )
-      && ( callError.includes("must be admin to call this function") )
-    );
-    test.log(`admins: [${admins}]`);
-  });
-  workspace.test('remove admins by admin', async (test, {alice, bob, multicall, root}) => {
+  workspace.test('remove admins by admin', async (test, {alice, multicall, root}) => {
     let callError: any;
     /**
      * Three cases to test here:
