+++ __tests__/multicall.ava.ts
@@ -10,27 +10,6 @@ export function tests(workspace: Workspace) {
   /**
    * call multicall
    */
-  workspace.test('multicall by non-admin', async (test, {alice, bob, multicall, root}) => {
-    let callError: any;
-    // bob isn't admin so he can't call multicall
-    try {
-      // try catch bacause contract should panick
-      await bob.call(
-        multicall.accountId,
-        'multicall',
-        {calls: [ [] ]},
-        {
-          gas: Gas.parse('5 Tgas'),
-          attachedDeposit: NEAR.from('1') // 1 yocto
-        }
-      );
-    } catch (error) { callError = getFunctionCallError(error) }
-    
-    test.true(
-      callError.includes("must be admin to call this function")
-    );
-    test.log(`type: "${callError.type}"`);
-  });
   workspace.test('multicall by admin', async (test, {alice, bob, multicall, testHelper, root}) => {
     // alice is admin so she can call multicall
     /**
