+++ __tests__/nearAPI.ava.ts
@@ -10,37 +10,6 @@ export function tests(workspace: Workspace) {
   /**
    * transfer NEAR
    */
-  workspace.test('transfer near by non-admin', async (test, {alice, bob, multicall, root}) => {
-    let callError: any;
-    const oldBalance = await multicall.balance();
-    test.log(`initial balance: ${oldBalance.total.toHuman()}`);
-    // bob isn't admin so he can't transfer near from multicall
-    try {
-      // try catch bacause contract should panick
-      await bob.call(
-        multicall.accountId,
-        'near_transfer',
-        {
-          account_id: bob.accountId,
-          amount: NEAR.parse("1").toString() // 1 NEAR
-        },
-        {
-          gas: Gas.parse('5 Tgas'),
-          attachedDeposit: NEAR.from('1') // 1 yocto
-        }
-      );
-    } catch (error) { callError = getFunctionCallError(error) }
-
-    const balance  = await multicall.balance();
-
-    test.true(
-      // balance won't decrease. Only increase by 30% of gas fees
-      // make sure balance did not decrease
-      ( ( oldBalance.total.sub( balance.total ) ).lte( NEAR.from("0") ) )
-      && ( callError.includes("must be admin to call this function") )
-    );
-    test.log(`balance: ${balance.total.toHuman()}`);
-  });
   workspace.test('transfer near by admin', async (test, {alice, bob, multicall, root}) => {
     const oldBalance = await multicall.balance();
     test.log(`initial balance: ${oldBalance.total.toHuman()}`);
