+++ __tests__/tokens.ava.ts
@@ -1,5 +1,5 @@
 import { Workspace, NEAR, Gas } from 'near-workspaces-ava';
-import { getFunctionCallError } from './helpers';
+
 
 const nusdc_address: string = "nusdc.ft-fin.testnet";
 const ndai_address: string = "ndai.ft-fin.testnet";
@@ -10,33 +10,7 @@ export function tests(workspace: Workspace) {
   /**
    * add tokens to whitelist
    */
-  workspace.test('add tokens to whitelist by non-admin', async (test, {alice, bob, multicall, root}) => {
-    let callError: any;
-    // bob isn't admin so he can't modify tokens whitelist
-    try {
-      // try catch bacause contract should panick
-      await bob.call(
-        multicall.accountId,
-        'tokens_add',
-        {
-          addresses: [nusdc_address]
-        },
-        {
-          gas: Gas.parse('5 Tgas'),
-          attachedDeposit: NEAR.from('1') // 1 yocto
-        }
-      );
-    } catch (error) { callError = getFunctionCallError(error) }
-
-    const tokens: string[] = await multicall.view('get_tokens', {})
-
-    test.true(
-      ( tokens.includes(ndai_address) && !tokens.includes(nusdc_address) )
-      && ( callError.includes("must be admin to call this function") )
-    );
-    test.log(`tokens: [${tokens}]`);
-  });
-  workspace.test('add tokens to whitelist by admin', async (test, {alice, bob, multicall, root}) => {
+  workspace.test('add tokens to whitelist by admin', async (test, {alice, multicall, root}) => {
     // alice is admin so she can modify tokens whitelist
     await alice.call(
       multicall.accountId,
@@ -59,32 +33,7 @@ export function tests(workspace: Workspace) {
   /**
    * remove tokens from whitelist
    */
-  workspace.test('remove tokens from whitelist by non-admin', async (test, {alice, bob, multicall, root}) => {
-    let callError: any;
-    // bob isn't admin so he can't modify tokens whitelist
-    try {
-      // try catch bacause contract should panick
-      await bob.call(
-        multicall.accountId,
-        'tokens_remove',
-        {
-          addresses: [ndai_address]
-        },
-        {
-          gas: Gas.parse('5 Tgas'),
-          attachedDeposit: NEAR.from('1') // 1 yocto
-        }
-      );
-    } catch (error) { callError = getFunctionCallError(error) }
-
-    const tokens: string[] = await multicall.view('get_tokens', {})
-    test.true(
-      ( tokens.includes(ndai_address) )
-      && ( callError.includes("must be admin to call this function") )
-    );
-    test.log(`tokens: [${tokens}]`);
-  });
-  workspace.test('remove tokens from whitelist by admin', async (test, {alice, bob, multicall, root}) => {
+  workspace.test('remove tokens from whitelist by admin', async (test, {alice, multicall, root}) => {
     // alice is admin so she can modify tokens whitelist
     await alice.call(
       multicall.accountId,
